+++ tests/testthat/test-netsim.R
@@ -24,6 +24,8 @@ for (trim in c(FALSE, TRUE)) {
     expect_is(mod, "netsim")
     plot(mod)
     plot(mod, type = "formation")
+    plot(mod, type = "duration")
+    plot(mod, type = "dissolution")    
     plot(mod, type = "network")
     test_net(mod)
   })
@@ -175,4 +177,33 @@ test_that("non-nested EDA works in netsim", {
                 coef.diss = dc, nested.edapprox = FALSE)
   dxs <- netdx(est, nsteps = 2, nsims = 2, dynamic = FALSE)
   sim <- netsim(est, param, init, control)
+})
+
+test_that("netsim diss.stats", {
+  nw <- network_initialize(n = 100)
+  formation <- ~edges
+  target.stats <- 50
+  coef.diss <- dissolution_coefs(dissolution = ~offset(edges), duration = 10)
+  est <- netest(nw, formation, target.stats, coef.diss, verbose = FALSE)
+
+  # Epidemic model
+  param <- param.net(inf.prob = 0.3)
+  init <- init.net(i.num = 10)
+  control <- control.net(type = "SI", nsteps = 5, nsims = 2, verbose = FALSE)
+  mod <- netsim(est, param, init, control)
+
+  print(mod)
+  expect_output(print(mod), "Dissolution Diagnostics.*Sim Mean")
+  expect_error(expect_output(print(mod), "Not available when:"))
+  expect_equal(length(mod[["diss.stats"]]), 2)
+  
+  mod2 <- merge(mod, mod)
+  expect_output(print(mod2), "Dissolution Diagnostics.*Sim Mean")
+  expect_error(expect_output(print(mod2), "Not available when:"))
+  expect_equal(length(mod2[["diss.stats"]]), 4)
+  
+  mod3 <- merge(mod, mod, keep.diss.stats = FALSE)
+  expect_error(expect_output(print(mod3), "Dissolution Diagnostics.*Sim Mean"))
+  expect_output(print(mod3), "Not available when:")
+  expect_true(is.null(mod3[["diss.stats"]]))
 })